backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==1.0.1
google-generativeai==0.3.2
chromadb==0.4.22
youtube-transcript-api==0.6.1
pytube==15.0.0
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.26.0
clerk-backend-api==1.0.0
sentence-transformers==2.2.2